delete the entire IDL system
