--- defs.ng
@@ -1,50 +1,0 @@
-package netmsg
-
-struct Multipart {
-    ID  uint16
-    GroupID uint32
-    NumParts uint16
-    Content []byte
-}
-
-struct Heartbeat {
-    Time int64
-    Latency int64
-}
-
-struct Benchy {
-	Name     string
-	BirthDay int64
-	Phone    string
-	Siblings int32
-	Spouse   byte
-	Money    float64
-}
-
-struct Nested {
-    A SubNest
-    B *SubNest2
-}
-
-struct ExampleDyn {
-  DynField dynamic
-}
-
-struct SubNest {
- B int32
- C float64
-}
-
-struct SubNest2 {
-
-}
-
-struct Connected {
-  Awesomeness Level
-}
-
-enum Level {
-    PrettyLow = 0
-    PrettyOk = 1
-    PrettyAwesome = 2
-}
